Extract permutation argument into a submodule
