+++ src/plonk/prover.rs
@@ -3,16 +3,15 @@ use std::iter;
 
 use super::{
     circuit::{Advice, Assignment, Circuit, Column, ConstraintSystem, Fixed},
-    Error, Proof, ProvingKey,
+    permutation, Error, Proof, ProvingKey,
 };
 use crate::arithmetic::{
-    eval_polynomial, get_challenge_scalar, parallelize, BatchInvert, Challenge, Curve, CurveAffine,
-    FieldExt,
+    eval_polynomial, get_challenge_scalar, Challenge, Curve, CurveAffine, FieldExt,
 };
 use crate::poly::{
     commitment::{Blind, Params},
     multiopen::{self, ProverQuery},
-    LagrangeCoeff, Polynomial, Rotation,
+    LagrangeCoeff, Polynomial,
 };
 use crate::transcript::{Hasher, Transcript};
 
@@ -177,197 +176,46 @@ impl<C: CurveAffine> Proof<C> {
         // Sample x_1 challenge
         let x_1: C::Scalar = get_challenge_scalar(Challenge(transcript.squeeze().get_lower_128()));
 
-        // Compute permutation product polynomial commitment
-        let mut permutation_product_polys = vec![];
-        let mut permutation_product_cosets = vec![];
-        let mut permutation_product_cosets_inv = vec![];
-        let mut permutation_product_commitments_projective = vec![];
-        let mut permutation_product_blinds = vec![];
-
-        // Iterate over each permutation
-        let mut permutation_modified_advice = pk
-            .vk
-            .cs
-            .permutations
-            .iter()
-            .zip(pk.permutations.iter())
-            // Goal is to compute the products of fractions
-            //
-            // (p_j(\omega^i) + \delta^j \omega^i \beta + \gamma) /
-            // (p_j(\omega^i) + \beta s_j(\omega^i) + \gamma)
-            //
-            // where p_j(X) is the jth advice column in this permutation,
-            // and i is the ith row of the column.
-            .map(|(columns, permuted_values)| {
-                let mut modified_advice = vec![C::Scalar::one(); params.n as usize];
-
-                // Iterate over each column of the permutation
-                for (&column, permuted_column_values) in columns.iter().zip(permuted_values.iter())
-                {
-                    parallelize(&mut modified_advice, |modified_advice, start| {
-                        for ((modified_advice, advice_value), permuted_advice_value) in
-                            modified_advice
-                                .iter_mut()
-                                .zip(witness.advice[column.index()][start..].iter())
-                                .zip(permuted_column_values[start..].iter())
-                        {
-                            *modified_advice *=
-                                &(x_0 * permuted_advice_value + &x_1 + advice_value);
-                        }
-                    });
-                }
-
-                modified_advice
-            })
-            .collect::<Vec<_>>();
-
-        // Batch invert to obtain the denominators for the permutation product
-        // polynomials
-        permutation_modified_advice
-            .iter_mut()
-            .flat_map(|v| v.iter_mut())
-            .batch_invert();
-
-        for (columns, mut modified_advice) in pk
-            .vk
-            .cs
-            .permutations
-            .iter()
-            .zip(permutation_modified_advice.into_iter())
-        {
-            // Iterate over each column again, this time finishing the computation
-            // of the entire fraction by computing the numerators
-            let mut deltaomega = C::Scalar::one();
-            for &column in columns.iter() {
-                let omega = domain.get_omega();
-                parallelize(&mut modified_advice, |modified_advice, start| {
-                    let mut deltaomega = deltaomega * &omega.pow_vartime(&[start as u64, 0, 0, 0]);
-                    for (modified_advice, advice_value) in modified_advice
-                        .iter_mut()
-                        .zip(witness.advice[column.index()][start..].iter())
-                    {
-                        // Multiply by p_j(\omega^i) + \delta^j \omega^i \beta
-                        *modified_advice *= &(deltaomega * &x_0 + &x_1 + advice_value);
-                        deltaomega *= &omega;
-                    }
-                });
-                deltaomega *= &C::Scalar::DELTA;
-            }
-
-            // The modified_advice vector is a vector of products of fractions
-            // of the form
-            //
-            // (p_j(\omega^i) + \delta^j \omega^i \beta + \gamma) /
-            // (p_j(\omega^i) + \beta s_j(\omega^i) + \gamma)
-            //
-            // where i is the index into modified_advice, for the jth column in
-            // the permutation
-
-            // Compute the evaluations of the permutation product polynomial
-            // over our domain, starting with z[0] = 1
-            let mut z = vec![C::Scalar::one()];
-            for row in 1..(params.n as usize) {
-                let mut tmp = z[row - 1];
-
-                tmp *= &modified_advice[row];
-                z.push(tmp);
-            }
-            let z = domain.lagrange_from_vec(z);
-
-            let blind = Blind(C::Scalar::rand());
-
-            permutation_product_commitments_projective.push(params.commit_lagrange(&z, blind));
-            permutation_product_blinds.push(blind);
-            let z = domain.lagrange_to_coeff(z);
-            permutation_product_polys.push(z.clone());
-            permutation_product_cosets
-                .push(domain.coeff_to_extended(z.clone(), Rotation::default()));
-            permutation_product_cosets_inv.push(domain.coeff_to_extended(z, Rotation(-1)));
-        }
-        let mut permutation_product_commitments =
-            vec![C::zero(); permutation_product_commitments_projective.len()];
-        C::Projective::batch_to_affine(
-            &permutation_product_commitments_projective,
-            &mut permutation_product_commitments,
-        );
-        let permutation_product_commitments = permutation_product_commitments;
-        drop(permutation_product_commitments_projective);
-
-        // Hash each permutation product commitment
-        for c in &permutation_product_commitments {
-            transcript
-                .absorb_point(c)
-                .map_err(|_| Error::TranscriptError)?;
-        }
+        // Commit to permutations, if any.
+        let permutations = if !pk.vk.cs.permutations.is_empty() {
+            Some(permutation::Proof::commit(
+                params,
+                pk,
+                &witness.advice,
+                x_0,
+                x_1,
+                &mut transcript,
+            )?)
+        } else {
+            None
+        };
 
         // Obtain challenge for keeping all separate gates linearly independent
         let x_2: C::Scalar = get_challenge_scalar(Challenge(transcript.squeeze().get_lower_128()));
 
+        // Evaluate the h(X) polynomial's constraint system expressions for the permutation constraints, if any.
+        let (permutations, permutation_expressions) = permutations
+            .map(|p| p.construct(pk, &advice_cosets, x_0, x_1))
+            .transpose()?
+            .map(|(p, expressions)| (Some(p), Some(expressions)))
+            .unwrap_or_default();
+
         // Evaluate the h(X) polynomial's constraint system expressions for the constraints provided
-        let h_poly =
-            iter::empty()
-                // Custom constraints
-                .chain(meta.gates.iter().map(|poly| {
-                    poly.evaluate(
-                        &|index| pk.fixed_cosets[index].clone(),
-                        &|index| advice_cosets[index].clone(),
-                        &|index| aux_cosets[index].clone(),
-                        &|a, b| a + &b,
-                        &|a, b| a * &b,
-                        &|a, scalar| a * scalar,
-                    )
-                }))
-                // l_0(X) * (1 - z(X)) = 0
-                .chain(
-                    permutation_product_cosets
-                        .iter()
-                        .cloned()
-                        .map(|coset| Polynomial::one_minus(coset) * &pk.l0),
+        let h_poly = iter::empty()
+            // Custom constraints
+            .chain(meta.gates.iter().map(|poly| {
+                poly.evaluate(
+                    &|index| pk.fixed_cosets[index].clone(),
+                    &|index| advice_cosets[index].clone(),
+                    &|index| aux_cosets[index].clone(),
+                    &|a, b| a + &b,
+                    &|a, b| a * &b,
+                    &|a, scalar| a * scalar,
                 )
-                // z(X) \prod (p(X) + \beta s_i(X) + \gamma) - z(omega^{-1} X) \prod (p(X) + \delta^i \beta X + \gamma)
-                .chain(pk.vk.cs.permutations.iter().enumerate().map(
-                    |(permutation_index, columns)| {
-                        let mut left = permutation_product_cosets[permutation_index].clone();
-                        for (advice, permutation) in columns
-                            .iter()
-                            .map(|&column| {
-                                &advice_cosets[pk.vk.cs.get_advice_query_index(column, 0)]
-                            })
-                            .zip(pk.permutation_cosets[permutation_index].iter())
-                        {
-                            parallelize(&mut left, |left, start| {
-                                for ((left, advice), permutation) in left
-                                    .iter_mut()
-                                    .zip(advice[start..].iter())
-                                    .zip(permutation[start..].iter())
-                                {
-                                    *left *= &(*advice + &(x_0 * permutation) + &x_1);
-                                }
-                            });
-                        }
-
-                        let mut right = permutation_product_cosets_inv[permutation_index].clone();
-                        let mut current_delta = x_0 * &C::Scalar::ZETA;
-                        let step = domain.get_extended_omega();
-                        for advice in columns.iter().map(|&column| {
-                            &advice_cosets[pk.vk.cs.get_advice_query_index(column, 0)]
-                        }) {
-                            parallelize(&mut right, move |right, start| {
-                                let mut beta_term =
-                                    current_delta * &step.pow_vartime(&[start as u64, 0, 0, 0]);
-                                for (right, advice) in right.iter_mut().zip(advice[start..].iter())
-                                {
-                                    *right *= &(*advice + &beta_term + &x_1);
-                                    beta_term *= &step;
-                                }
-                            });
-                            current_delta *= &C::Scalar::DELTA;
-                        }
-
-                        left - &right
-                    },
-                ))
-                .fold(domain.empty_extended(), |h_poly, v| h_poly * x_2 + &v);
+            }))
+            // Permutation constraints, if any.
+            .chain(permutation_expressions.into_iter().flatten())
+            .fold(domain.empty_extended(), |h_poly, v| h_poly * x_2 + &v);
 
         // Divide by t(X) = X^{params.n} - 1.
         let h_poly = domain.divide_by_vanishing_poly(h_poly);
@@ -402,7 +250,6 @@ impl<C: CurveAffine> Proof<C> {
         }
 
         let x_3: C::Scalar = get_challenge_scalar(Challenge(transcript.squeeze().get_lower_128()));
-        let x_3_inv = domain.rotate_omega(x_3, Rotation(-1));
 
         // Evaluate polynomials at omega^i x_3
         let advice_evals: Vec<_> = meta
@@ -432,27 +279,6 @@ impl<C: CurveAffine> Proof<C> {
             })
             .collect();
 
-        let permutation_product_evals: Vec<C::Scalar> = permutation_product_polys
-            .iter()
-            .map(|poly| eval_polynomial(poly, x_3))
-            .collect();
-
-        let permutation_product_inv_evals: Vec<C::Scalar> = permutation_product_polys
-            .iter()
-            .map(|poly| eval_polynomial(poly, domain.rotate_omega(x_3, Rotation(-1))))
-            .collect();
-
-        let permutation_evals: Vec<Vec<C::Scalar>> = pk
-            .permutation_polys
-            .iter()
-            .map(|polys| {
-                polys
-                    .iter()
-                    .map(|poly| eval_polynomial(poly, x_3))
-                    .collect()
-            })
-            .collect();
-
         let h_evals: Vec<_> = h_pieces
             .iter()
             .map(|poly| eval_polynomial(poly, x_3))
@@ -464,12 +290,12 @@ impl<C: CurveAffine> Proof<C> {
             .chain(aux_evals.iter())
             .chain(fixed_evals.iter())
             .chain(h_evals.iter())
-            .chain(permutation_product_evals.iter())
-            .chain(permutation_product_inv_evals.iter())
-            .chain(permutation_evals.iter().flat_map(|evals| evals.iter()))
         {
             transcript.absorb_scalar(*eval);
         }
+
+        // Evaluate the permutations, if any, at omega^i x_3.
+        let permutations = permutations.map(|p| p.evaluate(pk, x_3, &mut transcript));
 
         let instances =
             iter::empty()
@@ -511,68 +337,23 @@ impl<C: CurveAffine> Proof<C> {
                         }),
                 );
 
-        // Handle permutation arguments, if any exist
-        let permutation_instances = if !pk.vk.cs.permutations.is_empty() {
-            Some(
-                iter::empty()
-                    // Open permutation product commitments at x_3
-                    .chain(
-                        permutation_product_polys
-                            .iter()
-                            .zip(permutation_product_blinds.iter())
-                            .zip(permutation_product_evals.iter())
-                            .map(|((poly, blind), eval)| ProverQuery {
-                                point: x_3,
-                                poly,
-                                blind: *blind,
-                                eval: *eval,
-                            }),
-                    )
-                    // Open permutation polynomial commitments at x_3
-                    .chain(
-                        pk.permutation_polys
-                            .iter()
-                            .zip(permutation_evals.iter())
-                            .flat_map(|(polys, evals)| polys.iter().zip(evals.iter()))
-                            .map(|(poly, eval)| ProverQuery {
-                                point: x_3,
-                                poly,
-                                blind: Blind::default(),
-                                eval: *eval,
-                            }),
-                    )
-                    // Open permutation product commitments at \omega^{-1} x_3
-                    .chain(
-                        permutation_product_polys
-                            .iter()
-                            .zip(permutation_product_blinds.iter())
-                            .zip(permutation_product_inv_evals.iter())
-                            .map(|((poly, blind), eval)| ProverQuery {
-                                point: x_3_inv,
-                                poly,
-                                blind: *blind,
-                                eval: *eval,
-                            }),
-                    ),
-            )
-        } else {
-            None
-        };
-
         let multiopening = multiopen::Proof::create(
             params,
             &mut transcript,
-            instances.chain(permutation_instances.into_iter().flatten()),
+            instances.chain(
+                permutations
+                    .as_ref()
+                    .map(|p| p.open(pk, x_3))
+                    .into_iter()
+                    .flatten(),
+            ),
         )
         .map_err(|_| Error::OpeningError)?;
 
         Ok(Proof {
             advice_commitments,
             h_commitments,
-            permutation_product_commitments,
-            permutation_product_evals,
-            permutation_product_inv_evals,
-            permutation_evals,
+            permutations: permutations.map(|p| p.build()),
             advice_evals,
             fixed_evals,
             aux_evals,
