+++ src/plonk/verifier.rs
@@ -6,7 +6,6 @@ use crate::arithmetic::{get_challenge_scalar, Challenge, CurveAffine, FieldExt};
 use crate::poly::{
     commitment::{Guard, Params, MSM},
     multiopen::VerifierQuery,
-    Rotation,
 };
 use crate::transcript::{Hasher, Transcript};
 
@@ -53,10 +52,8 @@ impl<'a, C: CurveAffine> Proof<C> {
         let x_1: C::Scalar = get_challenge_scalar(Challenge(transcript.squeeze().get_lower_128()));
 
         // Hash each permutation product commitment
-        for c in &self.permutation_product_commitments {
-            transcript
-                .absorb_point(c)
-                .map_err(|_| Error::TranscriptError)?;
+        if let Some(p) = &self.permutations {
+            p.absorb_commitments(&mut transcript)?;
         }
 
         // Sample x_2 challenge, which keeps the gates linearly independent.
@@ -72,7 +69,6 @@ impl<'a, C: CurveAffine> Proof<C> {
         // Sample x_3 challenge, which is used to ensure the circuit is
         // satisfied with high probability.
         let x_3: C::Scalar = get_challenge_scalar(Challenge(transcript.squeeze().get_lower_128()));
-        let x_3_inv = vk.domain.rotate_omega(x_3, Rotation(-1));
 
         // This check ensures the circuit is satisfied so long as the polynomial
         // commitments open to the correct values.
@@ -84,9 +80,13 @@ impl<'a, C: CurveAffine> Proof<C> {
             .chain(self.aux_evals.iter())
             .chain(self.fixed_evals.iter())
             .chain(self.h_evals.iter())
-            .chain(self.permutation_product_evals.iter())
-            .chain(self.permutation_product_inv_evals.iter())
-            .chain(self.permutation_evals.iter().flat_map(|evals| evals.iter()))
+            .chain(
+                self.permutations
+                    .as_ref()
+                    .map(|p| p.evals())
+                    .into_iter()
+                    .flatten(),
+            )
         {
             transcript.absorb_scalar(*eval);
         }
@@ -130,59 +130,19 @@ impl<'a, C: CurveAffine> Proof<C> {
                         }),
                 );
 
-        // Handle permutation arguments, if any exist
-        let permutation_queries = if !vk.cs.permutations.is_empty() {
-            Some(
-                iter::empty()
-                    // Open permutation product commitments at x_3
-                    .chain(
-                        self.permutation_product_commitments
-                            .iter()
-                            .enumerate()
-                            .zip(self.permutation_product_evals.iter())
-                            .map(|((idx, _), &eval)| VerifierQuery {
-                                point: x_3,
-                                commitment: &self.permutation_product_commitments[idx],
-                                eval,
-                            }),
-                    )
-                    // Open permutation commitments for each permutation argument at x_3
-                    .chain(
-                        (0..vk.permutation_commitments.len())
-                            .map(|outer_idx| {
-                                let inner_len = vk.permutation_commitments[outer_idx].len();
-                                (0..inner_len).map(move |inner_idx| VerifierQuery {
-                                    point: x_3,
-                                    commitment: &vk.permutation_commitments[outer_idx][inner_idx],
-                                    eval: self.permutation_evals[outer_idx][inner_idx],
-                                })
-                            })
-                            .flatten(),
-                    )
-                    // Open permutation product commitments at \omega^{-1} x_3
-                    .chain(
-                        self.permutation_product_commitments
-                            .iter()
-                            .enumerate()
-                            .zip(self.permutation_product_inv_evals.iter())
-                            .map(|((idx, _), &eval)| VerifierQuery {
-                                point: x_3_inv,
-                                commitment: &self.permutation_product_commitments[idx],
-                                eval,
-                            }),
-                    ),
-            )
-        } else {
-            None
-        };
-
         // We are now convinced the circuit is satisfied so long as the
         // polynomial commitments open to the correct values.
         self.multiopening
             .verify(
                 params,
                 &mut transcript,
-                queries.chain(permutation_queries.into_iter().flatten()),
+                queries.chain(
+                    self.permutations
+                        .as_ref()
+                        .map(|p| p.queries(vk, x_3))
+                        .into_iter()
+                        .flatten(),
+                ),
                 msm,
             )
             .map_err(|_| Error::OpeningError)
@@ -209,29 +169,10 @@ impl<'a, C: CurveAffine> Proof<C> {
             return Err(Error::IncompatibleParams);
         }
 
-        if self.permutation_evals.len() != vk.cs.permutations.len() {
-            return Err(Error::IncompatibleParams);
-        }
-
-        for (permutation_evals, permutation) in
-            self.permutation_evals.iter().zip(vk.cs.permutations.iter())
-        {
-            if permutation_evals.len() != permutation.len() {
-                return Err(Error::IncompatibleParams);
-            }
-        }
-
-        if self.permutation_product_inv_evals.len() != vk.cs.permutations.len() {
-            return Err(Error::IncompatibleParams);
-        }
-
-        if self.permutation_product_evals.len() != vk.cs.permutations.len() {
-            return Err(Error::IncompatibleParams);
-        }
-
-        if self.permutation_product_commitments.len() != vk.cs.permutations.len() {
-            return Err(Error::IncompatibleParams);
-        }
+        self.permutations
+            .as_ref()
+            .map(|p| p.check_lengths(vk))
+            .transpose()?;
 
         // TODO: check h_commitments
 
@@ -275,46 +216,12 @@ impl<'a, C: CurveAffine> Proof<C> {
                     &|a, scalar| a * &scalar,
                 )
             }))
-            // l_0(X) * (1 - z(X)) = 0
             .chain(
-                self.permutation_product_evals
-                    .iter()
-                    .map(|product_eval| l_0 * &(C::Scalar::one() - product_eval)),
-            )
-            // z(X) \prod (p(X) + \beta s_i(X) + \gamma)
-            // - z(omega^{-1} X) \prod (p(X) + \delta^i \beta X + \gamma)
-            .chain(
-                vk.cs
-                    .permutations
-                    .iter()
-                    .zip(self.permutation_evals.iter())
-                    .zip(self.permutation_product_evals.iter())
-                    .zip(self.permutation_product_inv_evals.iter())
-                    .map(
-                        |(((columns, permutation_evals), product_eval), product_inv_eval)| {
-                            let mut left = *product_eval;
-                            for (advice_eval, permutation_eval) in columns
-                                .iter()
-                                .map(|&column| {
-                                    self.advice_evals[vk.cs.get_advice_query_index(column, 0)]
-                                })
-                                .zip(permutation_evals.iter())
-                            {
-                                left *= &(advice_eval + &(x_0 * permutation_eval) + &x_1);
-                            }
-
-                            let mut right = *product_inv_eval;
-                            let mut current_delta = x_0 * &x_3;
-                            for advice_eval in columns.iter().map(|&column| {
-                                self.advice_evals[vk.cs.get_advice_query_index(column, 0)]
-                            }) {
-                                right *= &(advice_eval + &current_delta + &x_1);
-                                current_delta *= &C::Scalar::DELTA;
-                            }
-
-                            left - &right
-                        },
-                    ),
+                self.permutations
+                    .as_ref()
+                    .map(|p| p.expressions(vk, &self.advice_evals, l_0, x_0, x_1, x_3))
+                    .into_iter()
+                    .flatten(),
             )
             .fold(C::Scalar::zero(), |h_eval, v| h_eval * &x_2 + &v);
